tests: move deploy to juju fixture and drop test indirection
